chore(project-structure): moved files around
