+++ src/components/File/styles.js
