+++ src/index.js
@@ -1,17 +1,3 @@
-/**
- * Form components
- */
-export * from './Form.js'
-export * from './ManagedForm.js'
-
-/**
- * Input components
- */
-export { File } from './File.js'
-export { Radio } from './Radio.js'
-export { RadioGroup } from './RadioGroup.js'
-
-/**
- * helpers
- */
+export * from './components.js'
+export * from './mutators.js'
 export * from './validators.js'
