+++ src/mutators.js
@@ -1,0 +1,1 @@
+export { clear } from './mutators/clear.js'
