+++ src/mutators/clear.js
