+++ src/validators.js
@@ -1,1 +1,3 @@
-export * from './validators/required.js'
+export { composeValidators } from './validators/composeValidators.js'
+export { required } from './validators/required.js'
+export { validEmail } from './validators/validEmail.js'
