+++ stories/Conditional.stories.js
@@ -1,31 +1,65 @@
-import React, { Fragment } from 'react'
+import React from 'react'
 
 import { storiesOf } from '@storybook/react'
 
-import { ManagedForm, File } from '../src'
-import { Conditional } from '../src/helper/Conditional.js'
+import { Form, File, Conditional, clear } from '../src/index.js'
+import { ManagedForm } from '../src/ManagedForm.js'
 
-storiesOf('Conditional fields', module).add('Default', () => (
-    <ManagedForm onSubmit={v => console.log(JSON.stringify(v, null, 2))}>
-        {({ values }) => (
-            <Fragment>
-                <File name="file1" label="File one" />
+storiesOf('Conditional fields', module)
+    .add('Plain form', () => (
+        <Form
+            onSubmit={values => console.log(JSON.stringify(values, null, 2))}
+            mutators={{ clear }}
+        >
+            {({ onSubmit, values }) => (
+                <form onSubmit={onSubmit}>
+                    <File name="file1" label="File one" />
 
-                <Conditional
-                    show={!!values.file1}
-                    fieldsToClearOnHide={['file3', 'file4']}
-                >
-                    <File name="file3" label="File three" />
-                    <File name="file4" label="File four" />
-                    <File name="file5" label="File five" />
-                    <div>
-                        <p>
-                            Conditional blocks can contain random other elements
-                            too
-                        </p>
-                    </div>
-                </Conditional>
-            </Fragment>
-        )}
-    </ManagedForm>
-))
+                    <Conditional
+                        show={!!values.file1}
+                        fieldsToClearOnHide={['file3', 'file4']}
+                    >
+                        <File name="file3" label="File three" />
+                        <File name="file4" label="File four" />
+                        <File name="file5" label="File five" />
+                        <div>
+                            <p>
+                                Conditional blocks can contain random other
+                                elements too
+                            </p>
+                        </div>
+                    </Conditional>
+
+                    <button type="submit">Submit</button>
+                </form>
+            )}
+        </Form>
+    ))
+    .add('Managed form', () => (
+        <ManagedForm
+            onSubmit={values => console.log(JSON.stringify(values, null, 2))}
+        >
+            {({ onSubmit, values }) => (
+                <React.Fragment>
+                    <File name="file1" label="File one" />
+
+                    <Conditional
+                        show={!!values.file1}
+                        fieldsToClearOnHide={['file3', 'file4']}
+                    >
+                        <File name="file3" label="File three" />
+                        <File name="file4" label="File four" />
+                        <File name="file5" label="File five" />
+                        <div>
+                            <p>
+                                Conditional blocks can contain random other
+                                elements too
+                            </p>
+                        </div>
+                    </Conditional>
+
+                    {/* <button type="submit">Submit</button> */}
+                </React.Fragment>
+            )}
+        </ManagedForm>
+    ))
